Python/TradeRoot_Jumio/requirements.txt
aws-requests-auth==0.4.3
azure-core==1.25.1
azure-identity==1.10.0
azure-monitor-opentelemetry-exporter==1.0.0b7
cachetools==5.2.0
certifi==2022.6.15.2
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Deprecated==1.2.13
google-api-core==2.10.0
google-auth==2.11.0
googleapis-common-protos==1.56.4
idna==3.4
isodate==0.6.1
msal==1.18.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.1
opencensus==0.11.0
opencensus-context==0.1.3
opencensus-ext-azure==1.1.7
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
portalocker==2.5.1
protobuf==4.21.6
psutil==5.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
typing-extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
